refactor: rename package torrent to torrentparser
