+++ torrentparser/parser.go
@@ -1,4 +1,4 @@
-package torrent
+package torrentparser
 
 import (
 	"errors"
