+++ torrentparser/parser_test.go
@@ -1,4 +1,4 @@
-package torrent
+package torrentparser
 
 import (
 	"bytes"
